기능별 class 분리 (static 사용하지 않는 의도)
