+++ src/main/java/baseball/Constant.java
@@ -1,0 +1,14 @@
+package baseball;
+
+public final class Constant {
+
+    public static final int INPUT_LENGTH_BOUND = 3;
+    public static final int GAME_CONTINUE = 1;
+    public static final int GAME_FINISH = 2;
+    public static final int CARD_COUNT = 3;
+    public static final int USED_NUMBER = 1;
+    public static final int ALL_STRIKE = 3;
+    public static final int NO_USED_CARD = -1;
+    public static final int WRONG = -1;
+
+}
